Outlines working on objects tagged "Interactable"
Make sure each object has a mesh collider and the outline script attached as well if you want to be able to get an outline on it when you look at it and be able to pick it up in the future.

I also had lots of issues with the TV with the new way I was doing things but I got that working now too with music and outline.